formatting

+++ src/IKUSD.sol
@@ -19,7 +19,7 @@ interface IKUSD {
 
     function transferFrom(address from, address to, uint256 value) external returns (bool);
 
-    function mint(address to, uint amount) external;
+    function mint(address to, uint256 amount) external;
 
-    function burn(address from, uint amount) external;
+    function burn(address from, uint256 amount) external;
 }
